+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - fixed warnings in myems-web
 - fixed warnings in myems-api
 - fixed issue of pop message in myems-admin
+- fixed issue of meter import function in myems-api
+- fixed issue of virtual meter import function in myems-api
 ### Removed
 - removed tbl_energy_storage_containers_sensors from myems_system_db in database